src/PersonIdentificationSystem.Python/stream_processor/requirements.txt
opencv-python-headless>=4.11.0,<5.0.0
httpx>=0.28.0,<1.0.0
numpy>=2.2.0,<3.0.0
requests>=2.32.0,<3.0.0
aio-pika>=9.5.0,<10.0.0
python-dotenv>=1.0.0,<2.0.0
aiohttp>=3.11.0,<4.0.0
